refactor(IW4-TU6): remove mpsp module
