+++ codxe.vcxproj
@@ -136,7 +136,6 @@
       <ClCompile Include="src\game\iw4\mp_tu6\components\fastfiles.cpp" />
       <ClCompile Include="src\game\iw4\mp_tu6\components\g_client_fields.cpp" />
       <ClCompile Include="src\game\iw4\mp_tu6\components\g_scr_main.cpp" />
-      <ClCompile Include="src\game\iw4\mp_tu6\components\mpsp.cpp" />
       <ClCompile Include="src\game\iw4\mp_tu6\components\mr.cpp" />
       <ClCompile Include="src\game\iw4\mp_tu6\components\patches.cpp" />
       <ClCompile Include="src\game\iw4\mp_tu6\components\pm.cpp" />
@@ -266,7 +265,6 @@
       <ClInclude Include="src\game\iw4\mp_tu6\components\fastfiles.h" />
       <ClInclude Include="src\game\iw4\mp_tu6\components\g_client_fields.h" />
       <ClInclude Include="src\game\iw4\mp_tu6\components\g_scr_main.h" />
-      <ClInclude Include="src\game\iw4\mp_tu6\components\mpsp.h" />
       <ClInclude Include="src\game\iw4\mp_tu6\components\mr.h" />
       <ClInclude Include="src\game\iw4\mp_tu6\components\patches.h" />
       <ClInclude Include="src\game\iw4\mp_tu6\components\pm.h" />
