+++ src/game/iw4/mp_tu6/main.cpp
@@ -7,7 +7,6 @@
 #include "components/fastfiles.h"
 #include "components/g_client_fields.h"
 #include "components/g_scr_main.h"
-#include "components/mpsp.h"
 #include "components/mr.h"
 #include "components/patches.h"
 #include "components/pm.h"
@@ -32,7 +31,6 @@ IW4_MP_TU6_Plugin::IW4_MP_TU6_Plugin()
     RegisterModule(new console());
     RegisterModule(new g_client_fields());
     RegisterModule(new g_scr_main());
-    RegisterModule(new mpsp());
     RegisterModule(new MovementRecorder());
     RegisterModule(new patches());
     RegisterModule(new pm());
